correctly type HOC and create global types
