+++ react/typings/global.d.ts
@@ -1,0 +1,11 @@
+import { FC } from 'react'
+import { TachyonsScaleInput } from '../hooks/tachyons'
+
+declare global {
+  interface CssHandlesOptions {
+    migrationFrom?: string | string[]
+  }
+  type CssHandlesInput = readonly string[]
+  type ValueOf<T extends readonly any[]> = T[number]
+  type CssHandles<T extends CssHandlesInput> = Record<ValueOf<T>, string>
+}
