+++ react/withCssHandles.tsx
@@ -2,15 +2,10 @@ import React from 'react'
 import useCssHandles from './useCssHandles'
 import { ComponentType } from 'react'
 
-interface CssHandlesOptions {
-  migrationFrom?: string | string[]
-}
-type CssHandlesInput = readonly string[]
-
 const withCssHandles = <T extends CssHandlesInput, ComponentProps>(
   handles: T,
   options?: CssHandlesOptions
-) => (Component: ComponentType<ComponentProps>) => {
+) => (Component: ComponentType<ComponentProps & { cssHandles: CssHandles<T> }>) => {
   const EnhancedComponent = (props: ComponentProps) => {
     const cssHandles = useCssHandles(handles, options)
